Pull in updates from master

+++ lib/str_util.cpp
@@ -860,8 +860,8 @@ vector<string> split(string s, char delim) {
 // - can't have ..
 //
 bool is_valid_filename(const char* name) {
-    int n = strlen(name);
-    for (int i=0; i<n; i++) {
+    size_t n = strlen(name);
+    for (size_t i=0; i<n; i++) {
         if (iscntrl(name[i])) {
             return false;
         }
